+++ src/events/log-file-path.ts
@@ -1,0 +1,1 @@
+export const logFilePath = './cypress/movie-events.log'
